serviço de alterar curso
- incluso o serviço que permite alterar curso
- Hotfix: correção do tipo do atributo rating

+++ src/main/java/com/coursemanager/domain/dto/CursoDTO.java
@@ -21,7 +21,7 @@ public class CursoDTO {
 	
 	private float duration;
 	
-	private short rating;
+	private float rating;
     
 	private Date releaseDate;
